organize IndexView
